Adds C++11 support

+++ daq/eventbuilder/CFragmentHandler.cpp
@@ -1266,7 +1266,7 @@ CFragmentHandler::generateBarrier(std::vector<EVB::pFragment>& outputList)
 	outputList.push_back(pFront);
 	p->second.s_queue.pop();
 	result.s_typesPresent.push_back(
-            std::pair<uint32_t, uint32_t>(p->first, pFront->s_header.s_barrier)
+            std::pair<uint32_t, uint32_t>(p->first, uint32_t(pFront->s_header.s_barrier))
         );
         m_nTotalFragmentSize -= pFront->s_header.s_size;
         if ((m_nTotalFragmentSize < m_nXonLimit)  && m_fXoffed) {
